Pull up 1.4, per request of fvdl.

+++ distrib/utils/sysinst/arch/arm32/md.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: md.c,v 1.1.2.3 1997/12/05 14:57:01 simonb Exp $	*/
+/*	$NetBSD: md.c,v 1.1.2.4 1998/05/29 04:04:38 mycroft Exp $	*/
 
 /*
  * Copyright 1997 Piermont Information Systems Inc.
@@ -248,11 +248,17 @@ void	md_post_newfs (void)
 
 void	md_copy_filesystem (void)
 {
+	if (target_already_root()) {
+		return;
+	}
+
 	/* Copy the instbin(s) to the disk */
 	printf ("%s", msg_string(MSG_dotar));
 	run_prog ("tar --one-file-system -cf - -C / . |"
 		  "(cd /mnt ; tar --unlink -xpf - )");
-	run_prog ("/bin/cp /tmp/.hdprofile /mnt/.profile");
+
+	/* Copy next-stage install profile into target /.profile. */
+	cp_to_target ("/tmp/.hdprofile", "/.profile");
 }
 
 int md_make_bsd_partitions (void)
